+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ### Changed
 * Errors instead of warns for opencv import issues
 * Warp methods now treat None as the identity transform.
+* `kwimage.num_channels` no longer errors when the number of channels is not 1, 3, or 4. 
 
 
 ## Version 0.7.2 - Released 2021-04-22